requirements.txt
# Web Framework
fastapi==0.111.0
uvicorn[standard]==0.30.1

# LLM - Anthropic Claude
anthropic==0.28.0

# Vector DB - pgvector + PostgreSQL
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy==2.0.31
alembic==1.13.2

# Document Ingestion & Parsing
pypdf==4.3.1
python-docx==1.1.2
python-pptx==0.6.23
unstructured==0.14.9
markdown==3.6

# Embeddings
sentence-transformers==3.0.1

# Text Chunking
langchain-text-splitters==0.2.2

# File Uploads
python-multipart==0.0.9

# Environment & Config
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4

# Utilities
tiktoken==0.7.0
tqdm==4.66.4

# Testing
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
